Implemented IO for output devices
This was kind of "tricky": since there was no output device in DAQCore, I implemented `OutputDev` that just stores the axes names and device name and config. And then when reading, if a more specific method is not implemented in other packages, it will assume that an `OutputDev` is used.